uradjeni servisi za korpu iteme korpe porudzbinu i iteme porudzbine, ispraviti servis porudzbine

+++ backend/hikeapp/src/main/java/hikeapp/dto/ProizvodDTO.java
@@ -1,5 +1,7 @@
 package hikeapp.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
@@ -30,7 +32,7 @@ public class ProizvodDTO {
 	
 	@NotNull(message = "Cena je obavezna")
 	@Positive(message = "Cena mora biti pozitivna")
-	private Double cena;
+	private BigDecimal cena;
 	
 	@Builder.Default
 	private Boolean dostupno = true;
